Moved login/register/profile links to MenuLinks.
